backend/pdf_converter/requirements.txt
# Minimal requirements for PDF converter Lambda
boto3==1.34.0
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.4.0

